Changed appID to be long instead of int. Appscanner uses app repository in some places instad of appscanningservice.

+++ src/SteamScrapper.Common/Utilities/Links/SteamLinkHelper.cs
@@ -85,7 +85,7 @@ namespace SteamScrapper.Common.Utilities.Links
             return appId;
         }
 
-        public static Uri CreateAppUri(int appId)
+        public static Uri CreateAppUri(long appId)
         {
             return new Uri(PageUrlPrefixes.App + appId.ToString(CultureInfo.InvariantCulture), UriKind.Absolute);
         }
