+++ src/SteamScrapper.Domain/Services/Contracts/AppData.cs
@@ -1,4 +1,4 @@
 ﻿namespace SteamScrapper.Domain.Services.Contracts
 {
-    public record AppData(int AppId, string Title, string BannerUrl);
+    public record AppData(long AppId, string Title, string BannerUrl);
 }
